Sepearate project properties from local properties
For all three templates the build process was modified,
so that local configuration for the build system can be seperated
from the project configuration.

This is required to allow several people to work on a library
or tool via version control.

Local configurations of the build system are now supposed to reside
in 'local.properties', a file which is excluded from revision control

Each template now has a 'local.properties.tmpl', that each developer
needs to customize and copy to 'local.properties'

For each of the three templates the following steps have been done:

 * create local.properties.tmpl
 * hide local.properties and distribution directory from revision control
 * update build process to parse both local.properties and build.properties
 * update changelog